mm: Add new vma flag VM_LOCAL_CPU
On a call to mmap an mmap provider (like an FS) can put
this flag on vma->vm_flags.

This tells the Kernel that the vma will be used from a single
core only and therefore invalidation of PTE(s) need not a
wide CPU scheduling

The motivation of this flag is the ZUFS project where we want
to optimally map user-application buffers into a user-mode-server
execute the operation and efficiently unmap.

In this project we utilize a per-core server thread so everything
is kept local. If we use the regular zap_ptes() API All CPU's
are scheduled for the unmap, though in our case we know that we
have only used a single core. The regular zap_ptes adds a very big
latency on every operations and mostly kills the concurrency of the
over all system. Because it imposes a serialization between all cores

Some preliminary measurements on a 40 core machines:

	unpatched		patched
Threads	Op/s	Lat [us]	Op/s	Lat [us]
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
1	185391	4.9		200799	4.6
2	197993	9.6		314321	5.9
4	310597	12.1		565574	6.6
8	546702	13.8		1113138	6.6
12	641728	17.2		1598451	6.8
18	744750	22.2		1648689	7.8
24	790805	28.3		1702285	8
36	849763	38.9		1783346	13.4
48	792000	44.6		1741873	17.4

[FIXME]
	We need to actually impose this policy. On very first
	pte_insert we should sample the used CPU_ID and on all
	susequent pte_inserts we need to make sure it is the
	same CPU_ID used.

NOTE: That this vma is never used during a page_fault. It is
always used in a synchronous way from an affinity set thread
to a single core.

Signed-off-by: Boaz Harrosh <boazh@netapp.com>