zuf: Write/Read && mmap implementation
The despatch to the server can operate on buffers up to
4 Mega bytes. Any bigger operations are split up and despatched
at this size.

Also if a multy-segments aio is used each segment is despatched
on its own. (TODO this can be easily fixed with sg operations)

On write if any mmaped buffers changed, for example new
allocated holes do to this write or a previous mmaped COW
was written. A range subset of the written range can be returned
for the Kernel to call mapping_unmap on.

mmap is achived with the GET_BLOCK operation. GET_BLOCK will
return if we need to unmap the previous mapped PTE in case
we are writing a prevouse faulted hole or in case of a COW.

Signed-off-by: Boaz Harrosh <boazh@netapp.com>